LU-3321 clio: revert LU-2622 for removing global env list
Revert "LU-2622 obdclass: Remove the global cl_env list"

This reverts commit d5390bbde9b525b13ab91f5610abe1c4bb98eff9.

The performance dropped a lot when memory reclaim process kicked
in as ll_releasepage() was called to destroy lustre pages. It turned
out that big overhead to allocate cl_env and keys on the fly so we
have to revert this patch.

The problem of LU-2622 will be solved in another patch later on.

Lustre-commit: 93fe562c5dd3829939e3bc0533918f66b19776a4
Lustre-change: http://review.whamcloud.com/7888

Signed-off-by: Jinshan Xiong <jinshan.xiong@intel.com>
Signed-off-by: Bob Glossman <bob.glossman@intel.com>
Reviewed-by: Niu Yawei <yawei.niu@intel.com>
Reviewed-by: Lai Siyao <lai.siyao@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>
Change-Id: Ib71f4808b9d9ca3e227c4d17eee3df2fdaadec69
Reviewed-on: http://review.whamcloud.com/9118
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>