Create new devel subpackage containing Lustre headers and libraries (lustre:13408 and lustre:13762 for Cray-master)
This is a port of lustre:13408 and lustre:13762 for Cray-master.

lustre:13408 created a new subpackage containing the Lustre headers and libraries
needed for development purposes. The subpackage is created for cray-lustre-ari_s and
cray-lustre-gem_s builds so that it can be made available on login nodes.

lustre:13762 fixed a problem where the headers were being added to the normal
packages in addition to the -devel subpackages. This resulted in breaking
the SUSE rpmlint checks for SP1.

Bugs Affected: 788353, 803583
Bug Action: UPDATE
Bug Comment:
Port for Cray-master
Test Description:
- Make new cray-lustre-cray_gem_s-devel and cray-lustre-cray_ari_s-devel builds.
- verify the content of the RPM builds.
- Install the RPM packages and verify the installed files on a centOS machine.
None


svn-id: lustre:14026