LU-3321 osc: add weight function for DLM lock
Use weigh_ast to decide if a lock covers any pages.
In recovery, weigh_ast will be used to decide if a DLM read lock
covers any locked pages, or it will be canceled instead being
recovered.

The problem with the original implementation is that it attached
each osc_page to an osc_lock also changed lock state to add every
pages for readahead.

Bugzilla-bug-Id: b=16774

Lustre-commit: 154fb1f7ce222a9b85cd8537aa1a0da9e25723fc
Lustre-change: http://review.whamcloud.com/7894

Signed-off-by: Jinshan Xiong <jinshan.xiong@intel.com>
Signed-off-by: Bob Glossman <bob.glossman@intel.com>
Reviewed-by: Bobi Jam <bobijam@gmail.com>
Reviewed-by: Lai Siyao <lai.siyao@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>
Change-Id: I760845b06d238e26bb056b85de901210c96db03a
Reviewed-on: http://review.whamcloud.com/9124
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>