Fix SP3 Lustre builds with OFED 3.5 (master)
This is merged from r14152: Fix SP3 Lustre b2_5 builds with OFED 3.5. The original description is:

OFED 3.5-2 supports the SLES 11 SP3 kernel.  CONFIG_COMPAT_SLES_11_3
macro is required to build lustre with this version of Lustre.
This is very similar to the change made upstream to support building
Lustre with OFED 3.5 with RHEL 6.4 (LU-3389).

This change requires looking at /etc/SuSE-release to determine whether
this is SLES 11 SP2 or SP3.  Adding a BuildRequires: sles-release  to
the appropriate Cray spec files ensures that it is present in the
build environment.

Bugs Affected: 804558
Bug Action: UPDATE
Test Description:
build successfully with master cray_gem_s
n/a

svn-id: lustre:14398