Remove cray-lustre-gem_s-sim.spec and all references to it (Cray-master)
Lustre builds for the gemini simulator are no longer supported so we no longer
need the cray-lustre-gem_s-sim.spec file.  This mod removes it and all
references to it.
Bugs Affected: 807030
Bug Action: UPDATE
Test Description:
Tested that other build still worked with changes to Makefile and obs config
NA


svn-id: lustre:14895