Build warnings from kgnilnd module that depend on gni kernel module for Apollo
lustre would not install properly on the apollos because it wasn't build with the gni Module.symvers.  While this only causes a warning on SLES, it stops the installation on CentOS.
Bugs Affected: 810916
Bug Action: RESOLVE
Test Description:
Ensured that the warnings for the Apollo build did not occur, and that the other flavors did not break.
No documentation affected

svn-id: lustre:15852