Change files statement in apollo specfiles.
The %_prefix is set to root, and all the root files are included by default.  %{dir} statements are not necessary (and exclude files that need to be included in the rpm), also, adding additional files statements (since we're already picking up the world).
Test Description:
Ensured the build compiled and rpm installed.
No documentation affected


svn-id: lustre:16085