Add module-init-tools requirement to rhine spec files
The rhine rpms call depmod in their %post and %postun scriptlets.  depmod is
included in the module-init-tools package, so this mod adds that package as a
requirement of the Rhine Lustre/LNET packages.

Also removed the requirement on %{switch_requires} (update-alternatives), since
we no longer call update-alternatives to set any default links because we aren't
doing release switching in Rhine.
Bugs Affected: 814355
Bug Action: UPDATE
NA

svn-id: lustre:17054