Add module-init-tools requirement to Rhine lnet subpackage
The "Requires: module-init-tools" in the preamble for the main Rhine lustre
package apparently did not apply to the lnet subpackage.  However, since the
lnet subpackage calls depmod in its %post and %postun scripts, it does depend on
module-init-tools.  This mod adds that "Requires: module-init-tools" tag to the
lnet subpackage preamble.
Bugs Affected: 814355
Bug Action: UPDATE
Test Description:
Built the rhine rpms and queried the dependencies.
NA

svn-id: lustre:17060