Create a Rhine spec file for Lustre/LNet on compute nodes
Create a spec file for building Lustre for compute nodes in CLE Rhine.  This is
a separate spec file because I want to isolate the changes to the Rhine compute
node Lustre from the Pearl compute node Lustre.  Currently these differences are
just a call to depmod in the %post and %postun scripts and a dependency on
module-init-tools (the package containing depmod).  I'm just being careful not
to break installs in Pearl with this change.
Bugs Affected: 814942
Bug Action: RESOLVE
Test Description:
Built and examined rpms with "rpm -ql", "rpm -q --requires", and "rpm -q --scripts".
NA

svn-id: lustre:17065