LU-4856 misc: Reduce exposure to overflow on page counters.
When the number of an object in use or circulation is tied to memory
size of the system, very large memory systems can overflow 32 bit
counters.  This patch addresses overflow on page counters in the osc LRU
and obd accounting.

This patch is back-ported from the following one:
Lustre-commit: e12bbee8207c41b2be5c8bb848f03bb81e75b025
Lustre-change: http://review.whamcloud.com/10537

Change-Id: Iffef7ec837fa9c94686fa0279079a75e803ccc45
Signed-off-by: Stephen Champion <schamp@sgi.com>
Signed-off-by: Jian Yu <jian.yu@intel.com>
Reviewed-on: http://review.whamcloud.com/12106
Tested-by: Jenkins
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>