LU-5560 llite: basic support of SELinux in CLIO
Bring the ability to properly initiate security context
on SELinux-enabled client and store it on server side via
extended attribute.

Security context initialization is not atomic, but that would
require a wire protocol change to send security label in the
creation request.

Filter out security.selinux from xattr cache as it is
already cached in system slab.

This patch is back-ported from the following one:
Lustre-commit: 8a11cb6282cfbdc8617b809344e6a11223e86a38
Lustre-change: http://review.whamcloud.com/11648

Signed-off-by: Sebastien Buisson <sebastien.buisson@bull.net>
Signed-off-by: John L. Hammond <john.hammond@intel.com>
Change-Id: Ie6db97be6f78b0b33da2ae54883a4051a2576af5
Reviewed-by: Dmitry Eremin <dmitry.eremin@intel.com>
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>
Reviewed-on: http://review.whamcloud.com/16253
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Bob Glossman <bob.glossman@intel.com>
Reviewed-by: Bobi Jam <bobijam@hotmail.com>