LU-5626 ldiskfs: update non-htree dotdot in rename
In 2.4+, when renaming a directory, its old dotdot entry will
be removed firstly, then the new dotdot entry is inserted, and
ldiskfs tries to append FID-in-dirent to the new entry.
But the space for dotdot entry may not be enough to hold
the new dotdot with FID-in-dirent, such as an MDT device
restored from file-level backup, or a device upgraded from 1.8.

In that case, for non-HTree directories, the ".." entry
will be written in the next available space in the directory
block.  This is invalid, as the ".." entry must be the
second entry in the block.

The same bug was fixed for HTree directories in LU-2638.
As Fan Yong said then: we do not want to introduce
complex logic to handle directory data moving, instead, in
such case, ignore the FID-in-dirent for the new dotdot entry,
and just insert the new dotdot entry.

There is one known flaw: This patch, like the one for
LU-2638, skips the entire data section rather than just
the FID.  This could cause trouble if something else ever
uses this section with ".." entries.

Patch ported from:
Lustre-change: http://review.whamcloud.com/#/c/12927/
Lustre-change: http://review.whamcloud.com/#/c/12926/

Signed-off-by: Patrick Farrell <paf@cray.com>
Signed-off-by: Wang Shilong <wshilong@ddn.com>
Change-Id: Iad8ac68a7b2c3c29e7765f0eaf18b07557259c50
Reviewed-on: http://review.whamcloud.com/17007
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Bob Glossman <bob.glossman@intel.com>
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>