LU-3974 client: use standard dentry locking function
Over time the struct dentry reference count have change
its data structure. In the 3.11 kernel it was changed
to a struct lockref. To handle these cases in the past
Lustre implemented a d_refcount wrapper to handle all
these changes. At the same time for 3.11 a d_count()
wrapper was introduced. So this patch moved use to use
this d_count wrapper and if not present we emulate it.

Lustre-commit: 266a0e3f717dcf74f750667e859c9382ddb8b817
Lustre-change: http://review.whamcloud.com/7741

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Signed-off-by: Bob Glossman <bob.glossman@intel.com>
Reviewed-by: Peng Tao <bergwolf@gmail.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>
Change-Id: I9afcf57c40dc699a073ae110164984b9cb54554d