LU-4416 osd-ldiskfs: vfs structs use fully typed uid/gid
Since Linux 3.5, struct inode and iattr have used type kuid_t
and kgid_t for uids and gids. There was a compatibility mode
to avoid build failures with older code but it was removed in
3.14. Furthermode, SLES12 builds with UIDGID_STRICT_TYPE_CHECKS
enabled, and needs to use the shim interfaces to build.

This patch provides the shims for older releases without them.

Lustre-commit: 5ecb7fd0b0af45a94af94ddaaa4f4fe1e59162bf
Lustre-change: http://review.whamcloud.com/10161

Signed-off-by: Jeff Mahoney <jeffm@suse.com>
Signed-off-by: James Simmons <uja.ornl@gmail.com>
Signed-off-by: Bob Glossman <bob.glossman@intel.com>
Reviewed-by: Peng Tao <bergwolf@gmail.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>
Change-Id: I8b10ccc0b340c26722826353c0de4d3728743e4a