LU-5363 build: add ll_d_count wrapper to fix build problem
Create and use ll_d_count() wrapper to solve
build problem caused by linux #include changes.

This comes from

Lustre-commit: 3000a83dcd3c201f383c0507d33e7861fc695997
Lustre-change: http://review.whamcloud.com/11133

It was left out of the previous backport of this mod as
it wasn't needed in the strictly v2.5 based code.
It is needed now due to the many additional changes added
for new kernel support.

Signed-off-by: Bob Glossman <bob.glossman@intel.com>
Change-Id: I871a30655bf0f1deaa72ae98c097e8200a054da4