LU-5781 osc: osc_lock_weight endless loop fix
With huge number of pages to scan by osc_lock_weight() it is likely
CLP_GANG_RESCHED is returned from osc_page_gang_lookup() and the scan
will be repeated again from the start. To be sure that the scan is
progressing across those restarts, next scan should be started from
the last scanned page index plus one.

This is the b_ieel2_0 backport of:
Lustre-commit: 74dfa6c3f6111750c773e2484b65302026af6a53
Lustre-change: http://review.whamcloud.com/12362

Signed-off-by: Alexander Zarochentsev <alexander.zarochentsev@seagate.com>
Signed-off-by: Jinshan Xiong <jinshan.xiong@intel.com>
Signed-off-by: Bruno Faccini <bruno.faccini@intel.com>
Change-Id: I60775f2d8f688029d97ad4fe64f9b2698ed278c8
Reviewed-by: Bobi Jam <bobijam@hotmail.com>
Reviewed-by: James Simmons <uja.ornl@gmail.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>
Reviewed-on: http://review.whamcloud.com/13999
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>