LU-5771 ldiskfs: cleanup orphan inode in error path
Commit from upstream: 4538821993f4486c76090dfb377c60c0a0e71ba3

There were some error paths in ext4_delete_inode() which was not
dropping the inode from the orphan list.  This could lead to a BUG_ON
on umount when the orphan list is discovered to be non-empty.

This is back-ported from the following patch:
Lustre-commit: 2dc56b1132a1680d664e8093a33f5ce799865abb
Lustre-change: http://review.whamcloud.com/12349

Change-Id: I0b1c9824d86481f346f032a2a18c421c684ab18e
Signed-off-by: Wang Shilong <wshilong@ddn.com>
Reviewed-on: http://review.whamcloud.com/14076
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>