LU-6081 lib: don't ignore the return of read()
The return of read() must be checked on some platform, possibly with a
recent version of glibc. Otherwise this gives an "ignoring return
value of 'read'" warning, which breaks the build. It happens on Ubuntu
14.04.

Use the return of read() to the random entropy to fix the issue. This
doesn't do much for the entropy, but it doesn't hurt either.

This is back-ported from the following patch:
Lustre-commit: 432dc1efa71e56e364b07c9cd0d0ba1feffadce1
Lustre-change: http://review.whamcloud.com/13654

Signed-off-by: frank zago <fzago@cray.com>
Change-Id: Ibf7ccffa41de4320511d9f8149848273a7d49999
Reviewed-on: http://review.whamcloud.com/14133
Tested-by: Jenkins
Reviewed-by: Bob Glossman <bob.glossman@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>