LU-6038 osd-zfs: Avoid redefining KM_SLEEP
Due to some long overdue memory management cleanup in the ZoL kmem
implementation the definition of KM_SLEEP has changed.  This change
was expected to be transparent to consumers but it causes issues
for Lustre because it explicitly redefines KM_SLEEP.  This was
originally done to avoid overriding the Linux slab interfaces.

This change implements a more portable fix.  Instead of preventing
the inclusion of the kmem.h header by setting the guard.  The
kmem_cache_* preprocessor macros are explictly undefined to make
the Linux interface available.

The related ZoL pull requests are as follows:

  https://github.com/zfsonlinux/spl/pull/414
  https://github.com/zfsonlinux/zfs/pull/2918

This is back-ported from the following patch:
Lustre-commit: 5a758a40de3b6b4e0aeb082195f61d86536233d2
Lustre-change: http://review.whamcloud.com/13096

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>
Change-Id: I53a1b4bca5c424a5be41743cc49f6ceb6c7a0357
Reviewed-on: http://review.whamcloud.com/14262
Reviewed-by: Bob Glossman <bob.glossman@intel.com>
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Nathaniel Clark <nathaniel.l.clark@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>