LU-6370 osc: disable to control per-OSC LRU budget
It will produce a problem where a system has lots of OSTs but
only a few of them are involved in doing read on the client side.
In that case, the per-OSC budget will cause the read ahead pages to
be dropped aggressively, which results in significantly performance
degradation.

This patch disabled per-OSC LRU budget. However, this is not a real
fix. A real fix should feed back memory pressure from LRU on the OSC
layer to the LLITE layer, so that read ahead algorithm can take that
information into consideration when it decides read ahead window size.

This patch is back-ported from the following one:
Lustre-commit: 642dd7e50b4ff39e91c5fd0a771a26c59b5b6637
Lustre-change: http://review.whamcloud.com/14347

Signed-off-by: Jinshan Xiong <jinshan.xiong@intel.com>
Change-Id: Idaa1c9c6e68da1c8c2a63ea8465009cb6a858a74
Reviewed-on: http://review.whamcloud.com/14419
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Bobi Jam <bobijam@hotmail.com>
Reviewed-by: James Simmons <uja.ornl@yahoo.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>