LU-6370 osc: disable to control per-OSC LRU budget
It will produce a problem where a system has lots of OSTs but
only a few of them are involved in doing read on the client side.
In that case, the per-OSC budget will cause the read ahead pages to
be dropped aggressively, which results in significantly performance
degradation.

This patch disabled per-OSC LRU budget. However, this is not a real
fix. A real fix should feed back memory pressure from LRU on the OSC
layer to the LLITE layer, so that read ahead algorithm can take that
information into consideration when it decides read ahead window size.

This patch is back-ported from the following one:
Lustre-commit: 642dd7e50b4ff39e91c5fd0a771a26c59b5b6637
Lustre-change: http://review.whamcloud.com/14347

Change-Id: Id326562d6d19e530a276f921764f3d0d4ed02c4e
Signed-off-by: Jinshan Xiong <jinshan.xiong@intel.com>
Reviewed-on: http://review.whamcloud.com/14593
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>