LU-6215 kernel: use f_path for inode dentry
Starting with linux kernel 2.6.17 the f_dentry field of
struct inode was moved to f_path from the top level. To
prevent breakage a backwards compatible macro was provided.
The linux kernel 3.18 removed that  macro which causes
lustre to fail to compile. This patch migrates the code
over to using the proper field.

Lustre-commit: 4a8f90cfdf8868d25f14b02214e51e7423d70160
Lustre-change: http://review.whamcloud.com/14598

Signed-off-by: James Simmons <uja.ornl@yahoo.com>
Signed-off-by: Bob Glossman <bob.glossman@intel.com>
Change-Id: I1497fa6667dfe1e72db0a6c7e2da78af9acdb1ee
Reviewed-by: Dmitry Eremin <dmitry.eremin@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>
Reviewed-on: http://review.whamcloud.com/17230
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jian Yu <jian.yu@intel.com>