LU-6654 osd: Add osd_iobuf concurrency control
The osd_iobuf struct in osd_thread_info is used by a thread
other than the one which owns the osd_thread_info.

The existing concurrency control (checking dr_numreqs) has
a race condition where dr_numreqs is decremented, which
opens the path for the other thread to eventually free
osd_thread_info, and then the first thread calls wakeup
on a wait queue which is inside the now freed
osd_thread_info.

This patch allocates iobuf separately from osd_thread_info
and uses a spinlock around the decrement & wakeup and the
freeing operation.  This means the freeing operation
cannot happen until the wakeup has occurred, as the freeing
thread must either sleep waiting for the decrement & wakeup,
or, if it does not sleep, the decrement has occurred, and
the spin lock is held until the wakeup is complete.

The spinlock then prevents the freeing until the wakeup is
complete.

A spinlock is used instead of moving the wait queue outside
of the iobuf because the wait queue is per-thread and the
freeing operation is part of the thread rolling up its
thread specific state.
Bugs Affected: 826472
Bug Action: RESOLVE
Test Description:
Tested in VM and on real hardware.  No
hangs/crashes running obdfilter-survey.
Jira: LU-6654
Gerrit: http://review.whamcloud.com/#/c/14963/
Patchset: PS1

svn-id: lustre:20003