LU-6766 build: add support for arm64
Expand lustre to support the ARM64 platform.

The architecture is reported as aarch64 so add that support to
lustre-build-linux.m4.

For the Red Hat/Fedora packaging, strip the aarch64 suffix for the
required kernel.

Add definitions for lstat_t, lstat_f, fstat_f and fstatat_f for aarch64
architecture in lustre/include/lustre/lustre_user.h.

Change-Id: I2d4a9e7be4d4c32ab99da802f1b2318764e4386e
Signed-off-by: Arnaud Guignard <arnaud.guignard@cea.fr>
Reviewed-on: http://review.whamcloud.com/15395
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: James Simmons <uja.ornl@yahoo.com>
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: Dmitry Eremin <dmitry.eremin@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>