Backport lustre:19607 to Cray-b2_7
This mod contains the non-gnilnd changes from lustre:19607.
We're backporting this mod to Cray-b2_7 so that we can land the changes
made to gnilnd by lustre:19607 into gnilnd/trunk.

The patch for 19607 applied cleanly, but there were additional uses
of cfs_socket_t and the libcfs_ipif* functions that I pruned manually.
These were in mxlnd.c and lnet/lnet/config.c

The change to gnilnd will be made in a separate mod to that project.

The description of lustre:19607 is as follow:
LU-6245 libcfs: remove tcpip abstraction from libcfs

Since libcfs no longer builds for user land we can
move the tcpip abstraction that exist to the LNET
layer which is the only place that uses it. Also
the migrated code will use native linux kernel
apis directly instead of with wrappers.
n/a


svn-id: lustre:20104