MRP-2680 tests: detect active facet based on current state
Lustre failover tests can not be ran under xperior control
on the setup with ${facet}_HOST != ${facet}failover_HOST
because of xperior does not care about facet state (active/passive).
t-f keeps this info in "${facet}active" files, which are created
when facet_failover() is executed first time in the test session.
Before facet_failover() executed these files are empty and
active facet is ${facet} by default.
In case when tests are executed test-by-test the active facet is
${facet}failover after 1st test completed, and 2nd test is started
having ${facet}failover active without this info stored in
${facet}active files.

Patch contains the following changes:
- add the active facet detection based on current lustre state;
- fix sanity-hsm defect: agt1_HOST is empty on CLIENTCOUNT=1 setup
because CLIENT1 is not set.

Signed-off-by: Elena Gryaznova <elena.gryaznova@seagate.com>
Change-Id: I90c5a15e9b83344b04e6a7c61a5a510e519d4822
Reviewed-on: http://es-gerrit.xyus.xyratex.com:8080/6668
Reviewed-by: Andriy Skulysh <andriy.skulysh@seagate.com>
Tested-by: Jenkins
Reviewed-by: Alexander Lezhoev <alexander.lezhoev@seagate.com>