Disable build of o2iblnd
lustre:19607 introduced a build regression when building o2iblnd. We
don't actually need the ko2iblnd.ko for CentOS DAL, so we'll workaround
the issue by just disabling the build of o2iblnd. LU-6763 was opened
to root cause the issue and fix it upstream

We may re-enable o2iblnd at some point in the future.
Bugs Affected: 827848
Bug Action: RESOLVE
n/a

svn-id: lustre:20111