LDEV-87 hsm: avoid leaking cars
In mdt_hsm_add_hal() if the object to be archived does not exist then
return an error. In mdt_hsm_agent_send() replace a spurious obj ==
NULL test with a test that obj == ERR_PTR(-ENOENT). In the same
function if the action list is not sent to the copytool then return
-ECANCELED.

Lustre-commit: 7326a2f33d44de960d2a9b61af8abb3ce5442dad
Lustre-change: http://review.whamcloud.com/15592

Signed-off-by: John L. Hammond <john.hammond@intel.com>
Signed-off-by: Bob Glossman <bob.glossman@intel.com>
Change-Id: Icfe74e87183bc5356d4c7627088b402805dcc164
Reviewed-by: Faccini Bruno <bruno.faccini@intel.com>
Reviewed-by: Jinshan Xiong <jinshan.xiong@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>
Reviewed-on: http://review.whamcloud.com/16411
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>