LU-6816 utils: remove libzfs_load_module() call
Remove libzfs_load_module() call from the OSD.  The function is no
longer exported by libzfs2 in ZFS master, as it was intended as an
internal function to begin with.  libzfs_init(), which is already
being used to set up the OSD, loads the module if necessary.

The intent of the call to libzfs_load_module apparently was to make
zfs_init() fail silently if ZFS was improperly installed (e.g. the
userspace libraries were missing), but fail with an error message if
ZFS initialization failed.  However it's not clear how this could ever
work.

Handling such error conditions will be dealt with in a separate ticket
as it's a larger concern.

Lustre-commit: 4b70236ecb9f51c49795d7c15819c9184cdc16e6
Lustre-change: http://review.whamcloud.com/15539

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>
Signed-off-by: Bob Glossman <bob.glossman@intel.com>
Change-Id: I97a2de7cd1a32c8251d805192aaf557f48840584
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: Nathaniel Clark <nathaniel.l.clark@intel.com>
Reviewed-by: Jinshan Xiong <jinshan.xiong@intel.com>
Reviewed-on: http://review.whamcloud.com/16459
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>