LU-5407 tests: Error message for replay-single 58b and 58c
In the replay-single test suite, tests 58b and 58c set an attribute,
retrieve the attribute and check that the original message matches
what was set. If large extended attributes is enabled, the attribute
written can be very long. In case of error, we only print the first
few charaters of the message that was set.

Also, call error if there is a problem when the client mounts the
file system on $MOUNT2.

This patch is back-ported from the following one:
Lustre-commit: 88e555dbabfc35521345851ff41516156217b1ec
Lustre-change: http://review.whamcloud.com/14766

Test-Parameters: alwaysuploadlogs testlist=replay-single
Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: Iccd5f9464d922f04425498356b169031b04aed28
Reviewed-on: http://review.whamcloud.com/15565
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Cliff White <cliff.white@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>