Remove OBD flag for striped directory support
In case someone with a non-Cray client tries to use DNE 2,
we need to remove the OBD_CONNECT flag for striped
directory support.

We remove it from both clients and servers, just in case
someone tries to use a Cray 2.7 client with a non-Cray
server that supports DNE 2.

This is in addition to the changes to lfs on Cray clients,
where lfs itself prevents creating striped directories and
returns a clear error message.  (So it's probably not
necessary on the clients, but seems like a good idea anyway.)

With the flag gone, attempts to create striped directories
from a client which has striped directory support will
fail with -ENOTSUPP (-524).  This is the best we can do
for non-Cray clients.
Bugs Affected: 827666
Bug Action: UPDATE
Test Description:
Tested on VMs, installed on server & tested
normal client.  Verified -ENOTSUPP for making striped directories.

svn-id: lustre:20489