LU-6356 ptlrpc: do not switch out-of-date context
When trying to replace a dead sec context with a new one, we must
ensure the new context is already up-to-date.
If it is not the case, just return from sptlrpc_req_replace_dead_ctx()
and come later when the new context has been updated.

This patch is back-ported from the following one:
Lustre-commit: 09e5f8ad23a880cb12fdf521b46a30e1f0a11d03
Lustre-change: http://review.whamcloud.com/15708

Signed-off-by: Sebastien Buisson <sebastien.buisson@bull.net>
Signed-off-by: John L. Hammond <john.hammond@intel.com>
Change-Id: I3b0862f838b917b568e08d46a24581a39c9459c1
Reviewed-by: Mike Pershin <mike.pershin@intel.com>
Reviewed-by: John L. Hammond <john.hammond@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>
Reviewed-on: http://review.whamcloud.com/16265
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Dmitry Eremin <dmitry.eremin@intel.com>