Client workaround for asynchronous flock support (2.7)
We make the client reject the non-failed '0' mode locks sent from a server
with asynchronous flock support (MRP-656/LU-2525). This feature
is found only on Seagate servers, and was crashing Cray clients.

Note that this patch differs from the 2.5 version of the patch
because LU-2177 (a required supporting mod) is already in 2.7.
This patch is just the Cray workaround to keep clients from
crashing.

The crash and fix are as follows:
When a CP_CALLBACK is received with a zero mode, the server
(with LU-2525) is telling the client to destroy the lock,
rather than grant it.  This is done either for flock deadlocks
(LDLM_FL_FLOCK_DEADLOCK set) or (with the asynchronous flock
feature on the server) to unlock a 'waiting' lock.

We handle the CP_CALLBACK and grant the lock with the zero
mode, then wake up the thread which requested it, which is
waiting in ldlm_flock_completion_ast.

That function recognizes the 0 mode lock, and if FLOCK_DEADLOCK
is not set, it returns -EIO to ldlm_cli_enqueue_fini.
ldlm_cli_enqueue_fini then calls failed_lock_cleanup, and,
crucially, specifies the original lock mode for the request.
That mode is used to decref the lock (rather than the 0 mode
it was granted with).
Bugs Affected: 827722
Bug Action: UPDATE
Test Description:
Built in 2.7, tested extensively in 2.5.
Jira: CLSTR-4248

svn-id: lustre:20537