LU-5420 mgc: MGC should retry for invalid import
This is a port of LU-5420 to Cray-b2_7. In upstream master, this is a fix to a problem caused by LU-4913 which in turn is a fix for the problem of PH-1037. In Cray-b2_7, LU-4913 was committed but reverted to avoid the LU-5420 problem. In master, LU-4913 was never reverted before landing of LU-5420. This patch therfore includes the un-revert of LU-4913, and the LU-5420. There was also a minor conflict due to LU-5814 and is resolved in this patch.

This patch resolves PH-1037.

Here is the original description of LU-5420:

LU-5420 mgc: MGC should retry for invalid import

After http://review.whamcloud.com/#/c/9967/ is landed, mgc does
not wait the import connected(state = FULL), then enqueue and
retrieve config log, which will cause the mount process to fail,
especially if the mgc is shared by multiple targets.
So once mgc enqueue is failed, it will give another chance to
wait the import to recover, if the import comes back in time,
it will try to enqueue again. Otherwise it will use local config log.

Signed-off-by: Wang Di <di.wang@intel.com>
Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>

Test Description:
Tested on husk-esms1 and osts can all start after reformat.
Jira: LU-5420
Gerrit: http://review.whamcloud.com/#/c/11258/

svn-id: lustre:20636