LU-6215 llite: handle backing_dev_info removal
For kernels before 3.20 the struct address_space contained
a pointer to a struct backing_dev_info. This backing_dev_info
had to setup with the backing_dev_info that was contained
in the super block. Now in newer kernels that backing_dev_info
pointer has been removed from the address_space. If you want
to use the backing_dev_info you need to access it from the
super block data structure directly. Lustre upstream removed
all the backing_dev_info setup for the inode i_mapping which
is no longer needed. This patch makes Lustre handle the bdi
inode backing_dev_info only for per 3.20 kernels now. This
change was due to:

Linux-commit: 6bec0035286119eefc32a5b1102127e6a4032cb2

Another change in the 3.20 kernels is the removal of the mmap
functionaly for the backing device. This was removed in with

Linux-commit: b4caecd48005fbed3949dde6c1cb233142fd69e9

With this change the flag BDI_CAP_MAP_COPY was removed.

Signed-off-by: James Simmons <uja.ornl@yahoo.com>
Change-Id: Ic2cd9227c99e3f66001ea9470cdc6bfdbf4f1c2c
Reviewed-on: http://review.whamcloud.com/15182
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Bob Glossman <bob.glossman@intel.com>
Reviewed-by: Dmitry Eremin <dmitry.eremin@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>