Add lbug_on_eviction option
Add an lbug_on_eviction proc interface.  When it is set to a non-zero value, it
will cause a node to LBUG on eviction.  On the client, this means whenever it is
evicted by a server.  On the server, this means whenever it evicts a client due
to expired locks.  The server will not LBUG when it evicts a client for ping
timeouts or when a user requests client eviction through the evict_client proc
interface.
Bugs Affected: 785175
Bug Action: UPDATE
Test Description:
Tested manually evicting the client from the MDS and OSS.

Without bug_on_eviction set, the client is evicted and reconnects.

With lbug_on_eviction set to 1, the client LBUGs on eviction.

Tested setting lbug_on_eviction to 1 on the OSS and the MDS, and setting
fail_loc=0x305 on the client while doing I/O. The OSS LBUG when they evict
the client for not responding to LDLM callback.
Jira: LU-5026
Test Notes:
Testing done in CentOS 6.5 vms on sdmvm.

svn-id: lustre:20792