LUS-2591 lprocfs: Add lbug_on_eviction option
Add an lbug_on_eviction proc interface.  When it is set to a non-zero
value on a client, it will cause the client to LBUG whenever it is
evicted by the server. Note, an MDS is a client to OSTs, and every
server is a client of MGS. Thus, it is probably desireable to leave
this set to zero on servers.

Signed-off-by: Chris Horn <hornc@cray.com>
Change-Id: Id9c956ab828fc74ab48dda4ab6e7289b4e7fd231
Reviewed-on: http://es-gerrit.dev.cray.com:8080/17259
Reviewed-by: Alexander Boyko <c17825@cray.com>
Reviewed-by: Alexander Zarochentsev <c17826@cray.com>
Tested-by: Elena Gryaznova <c17455@cray.com>