LUS-5724 llite: Revert to single page mds_readpage RPC size
Lustre 2.x added support for larger readdir RPCs (mds_readpage) [from
LU-5]. The larger RPC responses take longer for the MDS to process,
and the length worsens with a workload of concurrent unlinks.  That
is because there is more opportunity for lock conflict (ping-pong)
for update of the directory entries. Clients spend more time waiting
for readdir responses while servers acquire exclusive directory locks
for the updates.  Also, clients drop all of their cached entries for
the unlink, which causes them to issue another unnecessary (large)
readdir.  POSIX states that readdir processing does not need to be
coherent with respect to directory updates after opendir (until
rewinddir is called).  Therefore, a solution would be to keep the
cached entries.  However, a fix for that would be more involved.
This change simply reverts the clients to return to a single page
mds_readpage RPC.  However, we want to retain whatever speedup the
large readdir RPCs could provide on interactive (login) nodes.
Therefore, we only disable large readdir support on compute nodes.

Signed-off-by: Chris Horn <hornc@cray.com>
Change-Id: I7d16f0d96aaea6c6ce72d57332523d52bd5a49a1
Reviewed-on: https://es-gerrit.dev.cray.com/17262
Reviewed-by: Justin Miller <jmiller@cray.com>
Reviewed-by: Alexander Boyko <c17825@cray.com>
Tested-by: Elena Gryaznova <c17455@cray.com>