Revert to single page mds_readpage RPC size (For Cray-b2_7)
This is a port of lustre:14029 for Cray-b2_7.

Lustre 2.x added support for larger readdir RPCs (mds_readpage) [from LU-5].
The larger RPC responses take longer for the MDS to process, and the length
worsens with a workload of concurrent unlinks.  That is because there is more
opportunity for lock conflict (ping-pong) for update of the directory entries.
Clients spend more time waiting for readdir responses while servers acquire
exclusive directory locks for the updates.  Also, clients drop all of their
cached entries for the unlink, which causes them to issue another unnecessary
(large) readdir.  POSIX states that readdir processing does not need to be
coherent with respect to directory updates after opendir (until rewinddir is
called).  Therefore, a solution would be to keep the cached entries.  However,
a fix for that would be more involved.  This change simply reverts the clients
to return to a single page mds_readpage RPC.  However, we want to retain
whatever speedup the large readdir RPCs could provide on interactive (login)
nodes.  Therefore, we only disable large readdir support on compute nodes.
Test Description:
Passed sanity tests using snake-p2 DAL.

svn-id: lustre:20839