preallocate all lprocfs_stats (For Cray-b2_7)
This is a port of lustre:14035 for Cray-b2_7.

Allocating the stats as-needed leads to fragmentation.

lprocfs_alloc_stats() preallocates only for IRQ_SAFE stats.  It does
this so that it doesn't have to try to allocate memory in an interrupt
handler, not for fragmentation reasons.

We remove the test for IRQ_SAFE so that we preallocate space for all
the stats.
Related to LU-1282, http://review.whamcloud.com/3246
Test Description:
Passed sanity tests on snake-p2 using DAL.

svn-id: lustre:20840